+++ src/js/Search/KeywordHistory.js
@@ -1,5 +1,5 @@
 import { CLASSNAME, LOCAL_STORAGE_KEY, NUMBER } from "../constants/index.js";
-import { $ } from "../utils/DOM.js";
+import { $ } from "../utils/index.js";
 import { messenger, MESSAGE } from "../messenger/index.js";
 
 export default class KeywordHistory {
