+++ src/js/Video/index.js
@@ -1,0 +1,14 @@
+import {
+  SEARCH_VIDEO_TEMPLATE,
+  WATCH_LATER_VIDEO_TEMPLATE,
+  HISTORY_VIDEO_TEMPLATE,
+} from "./template.js";
+import { renderMainVideo, renderSearchVideo } from "./render.js";
+
+export {
+  SEARCH_VIDEO_TEMPLATE,
+  WATCH_LATER_VIDEO_TEMPLATE,
+  HISTORY_VIDEO_TEMPLATE,
+  renderMainVideo,
+  renderSearchVideo,
+};
