+++ src/js/utils/index.js
@@ -1,0 +1,12 @@
+import { REDIRECT_SERVER_HOST, fetchYoutubeData } from "./API.js";
+import { $, $$ } from "./DOM.js";
+import { showSnackbar, showModalSnackbar } from "./snackbar.js";
+
+export {
+  REDIRECT_SERVER_HOST,
+  fetchYoutubeData,
+  $,
+  $$,
+  showSnackbar,
+  showModalSnackbar,
+};
